projects/compression/requirements.txt
amazon.ion
cbor2
lz4
msgpack
orjson
pandas
pyarrow
pymongo
python-rapidjson
simplejson
ujson
zstandard
